Revert "restructure jarvis"
