+++ Jarvis/src/jarvis_mcp/capabilities/models.py
