+++ pyproject.toml
@@ -32,14 +32,14 @@ chronolog-mcp = "chronomcp.server:main"
 compression-mcp = "server:main"
 darshan-mcp = "darshan_mcp.server:main"
 hdf5-mcp = "mcp_server.server:main"
-jarvis-mcp = "src.server:main"
+jarvis-mcp = "jarvis_mcp.server:main"
 lmod-mcp = "lmod_mcp.server:main"
-node-hardware-mcp = "src.server:main"
-pandas-mcp = "src.server:main"
+node-hardware-mcp = "mcp_server.server:main"
+pandas-mcp = "pandasmcp.server:main"
 parallel-sort-mcp = "server:main"
 parquet-mcp = "server:main"
-plot-mcp = "src.server:main"
-slurm-mcp = "src.server:main"
+plot-mcp = "plot.server:main"
+slurm-mcp = "slurm_mcp.server:main"
 
 [tool.setuptools]
 packages = []
